requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-celery-beat==2.5.0
django-import-export==3.2.0
django-timezone-field==5.0
et-xmlfile==1.1.0
flake8==6.0.0
getenv==0.2.0
idna==3.4
kombu==5.3.0
MarkupPy==1.14
mccabe==0.7.0
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.2
pep8==1.7.1
Pillow==10.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PySimpleGUI==4.60.5
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
